# specs/001-gastrovoice-ai-system/contracts/api.openapi.yaml
openapi: "3.1.0"
info:
  title: GastroVoice AI — Backend API
  version: "1.0.0"
  description: >
    REST API for the GastroVoice AI restaurant management system.
    Covers authentication, restaurant configuration, menu management,
    takeaway orders, reservations and voice assistant simulation.
servers:
  - url: http://localhost:8080/api
    description: Local Docker Compose
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:    { type: string }
        message: { type: string }
    UUID:
      type: string
      format: uuid

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    Reservation:
      type: object
      properties:
        id:            { $ref: "#/components/schemas/UUID" }
        reservationDate: { type: string, format: date }
        timeSlot:       { type: string, pattern: "^([01]\\d|2[0-3]):[0-5]\\d$" }
        numPeople:      { type: integer, minimum: 1 }
        customerName:   { type: string }
        customerPhone:  { type: string, nullable: true }
        customerEmail:  { type: string, format: email, nullable: true }
        status:         { type: string, enum: [pending, confirmed, cancelled, no_show] }
        notes:          { type: string, nullable: true }
        createdAt:      { type: string, format: date-time }
    CreateReservationRequest:
      type: object
      required: [reservationDate, timeSlot, numPeople, customerName, customerPhone]
      properties:
        reservationDate: { type: string, format: date }
        timeSlot:        { type: string, pattern: "^([01]\\d|2[0-3]):[0-5]\\d$" }
        numPeople:       { type: integer, minimum: 1 }
        customerName:    { type: string }
        customerPhone:   { type: string }
        customerEmail:   { type: string, format: email }
        notes:           { type: string }
    AvailabilityResponse:
      type: object
      properties:
        isAvailable:      { type: boolean }
        availableCapacity: { type: integer }
        nextAvailableSlot: { type: string, nullable: true, description: "ISO 8601 datetime" }

    # ── Voice ─────────────────────────────────────────────────────────────────
    VoiceSimulateRequest:
      type: object
      required: [text, restaurantId]
      properties:
        text:              { type: string, description: "User message text" }
        restaurantId:      { $ref: "#/components/schemas/UUID" }
        sessionId:         { type: string, nullable: true, description: "Existing session ID for multi-turn" }
        callerId:          { type: string, default: unknown }
        restaurantContext: { type: object, additionalProperties: true }
    VoiceSimulateResponse:
      type: object
      properties:
        sessionId: { type: string }
        intent:    { type: string }
        reply:     { type: string }
        data:      { type: object, additionalProperties: true }

# ─────────────────────────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /reservations/availability:
    get:
      tags: [reservations]
      summary: Check slot availability
      parameters:
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
        - name: timeSlot
          in: query
          required: true
          schema: { type: string }
        - name: numPeople
          in: query
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Availability result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AvailabilityResponse" }
  /reservations:
    get:
      tags: [reservations]
      summary: List reservations (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: date
          in: query
          schema: { type: string, format: date }
        - name: status
          in: query
          schema: { type: string, enum: [pending, confirmed, cancelled, no_show] }
      responses:
        "200":
          description: Reservation list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Reservation" }
    post:
      tags: [reservations]
      summary: Create a reservation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateReservationRequest" }
      responses:
        "201":
          description: Created reservation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Reservation" }
        "409":
          description: Slot full — reservation rejected
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /reservations/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { $ref: "#/components/schemas/UUID" }
    get:
      tags: [reservations]
      summary: Get reservation detail
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Reservation detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Reservation" }
  /reservations/{id}/cancel:
    post:
      tags: [reservations]
      summary: Cancel a reservation
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/UUID" }
      responses:
        "200":
          description: Cancelled reservation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Reservation" }

  # ── Voice ───────────────────────────────────────────────────────────────────
  /voice/simulate:
    post:
      tags: [voice]
      summary: Simulate a voice call turn via HTTP (no audio)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VoiceSimulateRequest" }
      responses:
        "200":
          description: Assistant reply and detected intent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VoiceSimulateResponse" }
  /voice/call:
    post:
      tags: [voice]
      summary: AGI audio callback from Asterisk
      description: >
        Receives a WAV audio file from the Asterisk AGI script, transcribes it,
        detects intent, executes business logic and returns TTS audio URL.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [sessionId, audio]
              properties:
                sessionId:    { type: string }
                restaurantId: { $ref: "#/components/schemas/UUID" }
                audio:        { type: string, format: binary, description: "WAV 8kHz 16-bit mono" }
      responses:
        "200":
          description: TTS audio URL and reply text
          content:
            application/json:
              schema:
                type: object
                properties:
                  audioUrl: { type: string, format: uri }
                  reply:    { type: string }
                  intent:   { type: string }